requirements.txt
absl-py==0.11.0
aliyun-python-sdk-core==2.13.26
astor==0.8.1
attrdict==2.0.1
attrs==19.3.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2016.2.28
cffi==1.14.5
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
ds-ctcdecoder==0.6.1
entrypoints==0.3
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy3==0.2.3
future==0.18.2
gast==0.4.0
gitdb==4.0.5
GitPython==3.1.14
google==3.0.0
grpcio==1.35.0
h5py==2.10.0
imageio==2.9.0
importlib-metadata==1.7.0
ipykernel==5.3.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
jiwer==2.1.0
jmespath==0.10.0
joblib==0.13.2
jsonpickle==1.5.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.31.0
lxml==4.6.2
Markdown==3.2.2
matplotlib==3.0.3
mistune==0.8.4
mock==3.0.5
munch==2.5.0
musdb==0.4.0
museval==0.3.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
notebook==6.0.3
numba==0.47.0
numexpr==2.7.2
numpy==1.16.5
opencv-python==4.3.0.38
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
portalocker==2.0.0
progressbar==2.5
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.12.4
py-cpuinfo==7.0.0
pyaml==20.4.0
pycparser==2.20
pycryptodome==3.9.8
pydub==0.24.1
Pygments==2.6.1
pyparsing==2.4.7
pypesq==1.2.4
pyrsistent==0.16.0
pystoi==0.3.3
python-dateutil==2.8.1
python-Levenshtein==0.12.0
python-speech-features==0.6
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
regex==2020.6.8
resampy==0.2.2
sacred==0.8.2
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
smmap==3.0.5
SoundFile==0.10.3.post1
soupsieve==2.1
sox==1.4.1
stempeg==0.2.3
tables==3.5.2
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==1.2.0
